Refactored tests for chars escaping
Including @jsumners test
